Rearrange chemistry data
