--- chemistryData.js
@@ -1,395 +1,0 @@
-// Periodic Table data
-export const periodicTable = {
-    H: 1.008,
-    He: 4.0026,
-    Li: 6.94,
-    Be: 9.0122,
-    B: 10.81,
-    C: 12.01,
-    N: 14.01,
-    O: 16.0,
-    F: 18.998,
-    Ne: 20.18,
-    Na: 22.99,
-    Mg: 24.305,
-    Al: 26.982,
-    Si: 28.085,
-    P: 30.974,
-    S: 32.07,
-    Cl: 35.45,
-    Ar: 39.948,
-    K: 39.098,
-    Ca: 40.078,
-    Sc: 44.956,
-    Ti: 47.867,
-    V: 50.942,
-    Cr: 51.996,
-    Mn: 54.938,
-    Fe: 55.845,
-    Co: 58.933,
-    Ni: 58.693,
-    Cu: 63.546,
-    Zn: 65.38,
-    Ga: 69.723,
-    Ge: 72.63,
-    As: 74.922,
-    Se: 78.971,
-    Br: 79.904,
-    Kr: 83.798,
-    Rb: 85.468,
-    Sr: 87.62,
-    Y: 88.906,
-    Zr: 91.224,
-    Nb: 92.906,
-    Mo: 95.95,
-    Tc: 98,
-    Ru: 101.07,
-    Rh: 102.91,
-    Pd: 106.42,
-    Ag: 107.87,
-    Cd: 112.41,
-    In: 114.82,
-    Sn: 118.71,
-    Sb: 121.76,
-    Te: 127.6,
-    I: 126.9,
-    Xe: 131.29,
-    Cs: 132.91,
-    Ba: 137.33,
-    La: 138.91,
-    Ce: 140.12,
-    Pr: 140.91,
-    Nd: 144.24,
-    Pm: 145,
-    Sm: 150.36,
-    Eu: 151.96,
-    Gd: 157.25,
-    Tb: 158.93,
-    Dy: 162.5,
-    Ho: 164.93,
-    Er: 167.26,
-    Tm: 168.93,
-    Yb: 173.05,
-    Lu: 174.97,
-    Hf: 178.49,
-    Ta: 180.95,
-    W: 183.84,
-    Re: 186.21,
-    Os: 190.23,
-    Ir: 192.22,
-    Pt: 195.08,
-    Au: 196.97,
-    Hg: 200.59,
-    Tl: 204.38,
-    Pb: 207.2,
-    Bi: 208.98,
-    Po: 209,
-    At: 210,
-    Rn: 222,
-    Fr: 223,
-    Ra: 226,
-    Ac: 227,
-    Th: 232.04,
-    Pa: 231.04,
-    U: 238.03,
-    Np: 237,
-    Pu: 244,
-    Am: 243,
-    Cm: 247,
-    Bk: 247,
-    Cf: 251,
-    Es: 252,
-    Fm: 257,
-    Md: 258,
-    No: 259,
-    Lr: 266,
-    Rf: 267,
-    Db: 270,
-    Sg: 271,
-    Bh: 270,
-    Hs: 277,
-    Mt: 278,
-    Ds: 281,
-    Rg: 282,
-    Cn: 285,
-    Nh: 286,
-    Fl: 289,
-    Mc: 290,
-    Lv: 293,
-    Ts: 294,
-    Og: 294,
-  };
-  
-  // Common hydrocarbon molecular weights
-  export const hydrocarbonWeights = {
-    CH4: 16.04, // Methane
-    C2H6: 30.07, // Ethane
-    C3H8: 44.1, // Propane
-    C4H10: 58.12, // Butane
-    C5H12: 72.15, // Pentane
-    C6H14: 86.18, // Hexane
-    C7H16: 100.21, // Heptane
-    C8H18: 114.22, // Octane
-    C9H20: 128.25, // Nonane
-    C10H22: 142.29, // Decane
-  
-    C2H2: 26.04, // Acetylene (Ethyne)
-    C2H4: 28.05, // Ethylene (Ethene)
-    C3H6: 42.08, // Propylene (Propene)
-    C4H8: 56.11, // Butylene (Butene)
-    C5H10: 70.13, // Pentene
-    C6H12: 84.16, // Hexene
-    C7H14: 98.19, // Heptene
-    C8H16: 112.21, // Octene
-    C9H18: 126.24, // Nonene
-    C10H20: 140.27, // Decene
-  
-    C6H6: 78.11, // Benzene
-    C7H8: 92.14, // Toluene
-    C8H10: 106.17, // Xylene
-    C9H12: 120.19, // Trimethylbenzene
-    C10H14: 134.22, // Tetramethylbenzene
-  
-    C6H10: 82.15, // Cyclohexene
-    C6H12: 84.16, // Cyclohexane
-    C7H14: 98.19, // Cycloheptane
-    C8H16: 112.21, // Cyclooctane
-  
-    C10H8: 128.17, // Naphthalene
-    C14H10: 178.23, // Anthracene
-    C14H10: 178.23, // Phenanthrene
-    C16H10: 202.26, // Pyrene
-    C18H12: 228.29, // Chrysene
-  
-    C20H12: 252.31, // Benzo[a]pyrene
-    C24H12: 300.36, // Coronene
-  };
-  
-  // Redox agents
-  export const oxidizingAgents = [
-    "O2", // Oxygen
-    "O3", // Ozone
-    "H2O2", // Hydrogen Peroxide
-    "KMnO4", // Potassium Permanganate
-    "HNO3", // Nitric Acid
-    "K2Cr2O7", // Potassium Dichromate
-    "CrO3", // Chromium Trioxide
-    "H2SO4", // Sulfuric Acid (concentrated)
-    "Cl2", // Chlorine
-    "Br2", // Bromine
-    "I2", // Iodine
-    "FeCl3", // Ferric Chloride
-    "CuSO4", // Copper(II) Sulfate
-    "MnO2", // Manganese Dioxide
-    "AgNO3", // Silver Nitrate
-    "NaOCl", // Sodium Hypochlorite (Bleach)
-    "PbO2", // Lead(IV) Oxide
-    "N2O", // Nitrous Oxide
-    "NO2", // Nitrogen Dioxide
-    "HClO", // Hypochlorous Acid
-    "ClO2", // Chlorine Dioxide
-    "HBrO", // Hypobromous Acid
-    "HIO3", // Iodic Acid
-    "XeF4", // Xenon Tetrafluoride
-    "C2H5NO2", // Glycine Nitrate
-  ];
-  
-  export const reducingAgents = [
-    "H2", // Hydrogen Gas
-    "C", // Carbon (Graphite)
-    "CO", // Carbon Monoxide
-    "CH4", // Methane
-    "C2H6", // Ethane
-    "C6H6", // Benzene
-    "Fe", // Iron
-    "Zn", // Zinc
-    "Al", // Aluminum
-    "Mg", // Magnesium
-    "Na", // Sodium
-    "K", // Potassium
-    "Li", // Lithium
-    "Ca", // Calcium
-    "Ba", // Barium
-    "Sn", // Tin
-    "Pb", // Lead
-    "NH3", // Ammonia
-    "H2S", // Hydrogen Sulfide
-    "NaBH4", // Sodium Borohydride
-    "LiAlH4", // Lithium Aluminum Hydride
-    "SO2", // Sulfur Dioxide
-    "FeSO4", // Iron(II) Sulfate
-    "TiCl3", // Titanium(III) Chloride
-    "Cu", // Copper
-    "Cr2O3", // Chromium(III) Oxide
-    "AsH3", // Arsine
-    "SbH3", // Stibine
-    "PH3", // Phosphine
-  ];
-  
-  // Common combustion fuels
-  export const validCombustionFuels = ["CH4", "C2H6", "C6H6"];
-  
-  // Acids and bases
-  export const acids = [
-    "HCl", // Hydrochloric acid
-    "H2SO4", // Sulfuric acid
-    "HNO3", // Nitric acid
-    "H3PO4", // Phosphoric acid
-    "CH3COOH", // Acetic acid
-    "H2CO3", // Carbonic acid
-    "HF", // Hydrofluoric acid
-    "HBr", // Hydrobromic acid
-    "HI", // Hydroiodic acid
-    "HClO", // Hypochlorous acid
-    "HClO2", // Chlorous acid
-    "HClO3", // Chloric acid
-    "HClO4", // Perchloric acid
-    "H2S", // Hydrosulfuric acid
-  ];
-  
-  export const bases = [
-    "NaOH", // Sodium hydroxide
-    "KOH", // Potassium hydroxide
-    "Ca(OH)2", // Calcium hydroxide
-    "Mg(OH)2", // Magnesium hydroxide
-    "NH3", // Ammonia
-    "Al(OH)3", // Aluminum hydroxide
-    "Fe(OH)3", // Iron (III) hydroxide
-    "NH4OH", // Ammonium hydroxide
-    "Ba(OH)2", // Barium hydroxide
-    "Sr(OH)2", // Strontium hydroxide
-    "LiOH", // Lithium hydroxide
-  ];
-  
-  // Physical constants
-  export const physicalConstants = {
-    R: 0.0821, // Ideal gas constant in L·atm/(mol·K)
-    standardTemperature: 273.15, // Kelvin
-    standardPressure: 1, // atm
-    molarVolume: 22.4, // L/mol at STP
-  };
-
-// Solubility table - maps anion + cation pairs to solubility status
-// P = soluble, H = insoluble, M = less soluble, "-" = does not exist or dissolves in water
-export const solubilityTable = {
-    // Format: "anion_cation": "solubility"
-    
-    // Hydroxide (OH-) solubility
-    "OH_H": "P", "OH_Li": "P", "OH_K": "P", "OH_Na": "P", "OH_NH4": "P",
-    "OH_Ba": "P", "OH_Ca": "M", "OH_Mg": "H", "OH_Al": "H", "OH_Cr": "H",
-    "OH_Fe2": "H", "OH_Fe3": "H", "OH_Ni": "H", "OH_Co": "H", "OH_Mn": "H",
-    "OH_Zn": "H", "OH_Ag": "-", "OH_Hg": "-", "OH_Rb": "H", "OH_Sn": "H", "OH_Cu": "H",
-    
-    // Fluoride (F-) solubility
-    "F_H": "P", "F_Li": "H", "F_K": "P", "F_Na": "P", "F_NH4": "P",
-    "F_Ba": "M", "F_Ca": "H", "F_Mg": "H", "F_Al": "P", "F_Cr": "P",
-    "F_Fe2": "M", "F_Fe3": "H", "F_Ni": "H", "F_Co": "M", "F_Mn": "M",
-    "F_Zn": "M", "F_Ag": "P", "F_Hg": "H", "F_Rb": "H", "F_Sn": "P", "F_Cu": "H",
-    
-    // Chloride (Cl-) solubility
-    "Cl_H": "P", "Cl_Li": "P", "Cl_K": "P", "Cl_Na": "P", "Cl_NH4": "P",
-    "Cl_Ba": "P", "Cl_Ca": "P", "Cl_Mg": "P", "Cl_Al": "P", "Cl_Cr": "P",
-    "Cl_Fe2": "P", "Cl_Fe3": "P", "Cl_Ni": "P", "Cl_Co": "P", "Cl_Mn": "P",
-    "Cl_Zn": "P", "Cl_Ag": "H", "Cl_Hg": "P", "Cl_Rb": "M", "Cl_Sn": "P", "Cl_Cu": "P",
-    
-    // Bromide (Br-) solubility
-    "Br_H": "P", "Br_Li": "P", "Br_K": "P", "Br_Na": "P", "Br_NH4": "P",
-    "Br_Ba": "P", "Br_Ca": "P", "Br_Mg": "P", "Br_Al": "P", "Br_Cr": "P",
-    "Br_Fe2": "P", "Br_Fe3": "P", "Br_Ni": "P", "Br_Co": "P", "Br_Mn": "P",
-    "Br_Zn": "P", "Br_Ag": "H", "Br_Hg": "M", "Br_Rb": "M", "Br_Sn": "P", "Br_Cu": "P",
-    
-    // Iodide (I-) solubility
-    "I_H": "P", "I_Li": "P", "I_K": "P", "I_Na": "P", "I_NH4": "P",
-    "I_Ba": "P", "I_Ca": "P", "I_Mg": "P", "I_Al": "P", "I_Cr": "P",
-    "I_Fe2": "P", "I_Fe3": "-", "I_Ni": "P", "I_Co": "P", "I_Mn": "P",
-    "I_Zn": "P", "I_Ag": "H", "I_Hg": "H", "I_Rb": "H", "I_Sn": "M", "I_Cu": "-",
-    
-    // Sulfide (S2-) solubility
-    "S_H": "P", "S_Li": "P", "S_K": "P", "S_Na": "P", "S_NH4": "P",
-    "S_Ba": "P", "S_Ca": "P", "S_Mg": "P", "S_Al": "-", "S_Cr": "-",
-    "S_Fe2": "H", "S_Fe3": "-", "S_Ni": "H", "S_Co": "H", "S_Mn": "H",
-    "S_Zn": "H", "S_Ag": "H", "S_Hg": "H", "S_Rb": "H", "S_Sn": "H", "S_Cu": "H",
-    
-    // Sulfate (SO4 2-) solubility
-    "SO4_H": "P", "SO4_Li": "P", "SO4_K": "P", "SO4_Na": "P", "SO4_NH4": "P",
-    "SO4_Ba": "H", "SO4_Ca": "H", "SO4_Mg": "H", "SO4_Al": "-", "SO4_Cr": "-",
-    "SO4_Fe2": "H", "SO4_Fe3": "-", "SO4_Ni": "H", "SO4_Co": "H", "SO4_Mn": "H",
-    "SO4_Zn": "H", "SO4_Ag": "H", "SO4_Hg": "-", "SO4_Rb": "H", "SO4_Sn": "-", "SO4_Cu": "-",
-    
-    // Sulfite (SO3 2-) solubility
-    "SO3_H": "P", "SO3_Li": "P", "SO3_K": "P", "SO3_Na": "P", "SO3_NH4": "P",
-    "SO3_Ba": "H", "SO3_Ca": "M", "SO3_Mg": "P", "SO3_Al": "P", "SO3_Cr": "P",
-    "SO3_Fe2": "P", "SO3_Fe3": "P", "SO3_Ni": "P", "SO3_Co": "P", "SO3_Mn": "P",
-    "SO3_Zn": "P", "SO3_Ag": "M", "SO3_Hg": "P", "SO3_Rb": "H", "SO3_Sn": "P", "SO3_Cu": "P",
-    
-    // Phosphate (PO4 3-) solubility
-    "PO4_H": "P", "PO4_Li": "H", "PO4_K": "P", "PO4_Na": "P", "PO4_NH4": "P",
-    "PO4_Ba": "H", "PO4_Ca": "H", "PO4_Mg": "H", "PO4_Al": "H", "PO4_Cr": "H",
-    "PO4_Fe2": "H", "PO4_Fe3": "H", "PO4_Ni": "H", "PO4_Co": "H", "PO4_Mn": "H",
-    "PO4_Zn": "H", "PO4_Ag": "H", "PO4_Hg": "H", "PO4_Rb": "H", "PO4_Sn": "H", "PO4_Cu": "H",
-    
-    // Carbonate (CO3 2-) solubility
-    "CO3_H": "P", "CO3_Li": "P", "CO3_K": "P", "CO3_Na": "P", "CO3_NH4": "P",
-    "CO3_Ba": "H", "CO3_Ca": "H", "CO3_Mg": "H", "CO3_Al": "-", "CO3_Cr": "-",
-    "CO3_Fe2": "H", "CO3_Fe3": "-", "CO3_Ni": "H", "CO3_Co": "H", "CO3_Mn": "H",
-    "CO3_Zn": "H", "CO3_Ag": "H", "CO3_Hg": "-", "CO3_Rb": "H", "CO3_Sn": "H", "CO3_Cu": "H",
-    
-    // Silicate (SiO3 2-) solubility
-    "SiO3_H": "H", "SiO3_Li": "P", "SiO3_K": "P", "SiO3_Na": "P", "SiO3_NH4": "-",
-    "SiO3_Ba": "H", "SiO3_Ca": "H", "SiO3_Mg": "H", "SiO3_Al": "H", "SiO3_Cr": "-",
-    "SiO3_Fe2": "H", "SiO3_Fe3": "H", "SiO3_Ni": "-", "SiO3_Co": "-", "SiO3_Mn": "H",
-    "SiO3_Zn": "H", "SiO3_Ag": "-", "SiO3_Hg": "-", "SiO3_Rb": "H", "SiO3_Sn": "-", "SiO3_Cu": "H",
-    
-    // Nitrate (NO3-) solubility
-    "NO3_H": "P", "NO3_Li": "P", "NO3_K": "P", "NO3_Na": "P", "NO3_NH4": "P",
-    "NO3_Ba": "P", "NO3_Ca": "P", "NO3_Mg": "P", "NO3_Al": "P", "NO3_Cr": "P",
-    "NO3_Fe2": "P", "NO3_Fe3": "P", "NO3_Ni": "P", "NO3_Co": "P", "NO3_Mn": "P",
-    "NO3_Zn": "P", "NO3_Ag": "P", "NO3_Hg": "P", "NO3_Rb": "P", "NO3_Sn": "-", "NO3_Cu": "P",
-    
-    // Acetate (CH3COO-) solubility
-    "CH3COO_H": "P", "CH3COO_Li": "P", "CH3COO_K": "P", "CH3COO_Na": "P", "CH3COO_NH4": "P",
-    "CH3COO_Ba": "P", "CH3COO_Ca": "P", "CH3COO_Mg": "P", "CH3COO_Al": "M", "CH3COO_Cr": "P",
-    "CH3COO_Fe2": "P", "CH3COO_Fe3": "P", "CH3COO_Ni": "P", "CH3COO_Co": "P", "CH3COO_Mn": "P",
-    "CH3COO_Zn": "P", "CH3COO_Ag": "P", "CH3COO_Hg": "P", "CH3COO_Rb": "P", "CH3COO_Sn": "P", "CH3COO_Cu": "P"
-  };
-
-
-// Function to extract cation and anion from a compound formula
-export function extractIons(formula) {
-    // we will need to improve it since this is just a headstart
-    // Common cations
-    const cations = {
-      "H": 1, "Li": 1, "K": 1, "Na": 1, "NH4": 1,
-      "Ba": 2, "Ca": 2, "Mg": 2, "Al": 3, "Cr": 3,
-      "Fe2": 2, "Fe3": 3, "Ni": 2, "Co": 2, "Mn": 2, 
-      "Zn": 2, "Ag": 1, "Hg": 2, "Rb": 1, "Sn": 2, "Cu": 2
-    };
-    
-    // Common anions
-    const anions = {
-      "OH": -1, "F": -1, "Cl": -1, "Br": -1, "I": -1,
-      "S": -2, "SO4": -2, "SO3": -2, "PO4": -3, "CO3": -2,
-      "SiO3": -2, "NO3": -1, "CH3COO": -1
-    };
-    
-    // // simplistic approach
-    // let cation = null;
-    // let anion = null;
-    
-    // // Try to identify the formula components
-    // for (const c in cations) {
-    //   if (formula.includes(c)) {
-    //     cation = c;
-    //     break;
-    //   }
-    // }
-    
-    // for (const a in anions) {
-    //   if (formula.includes(a)) {
-    //     anion = a;
-    //     break;
-    //   }
-    // }
-    
-    // return { cation, anion };
-  }
