+++ chemistryData/acidsBases.js
@@ -1,0 +1,30 @@
+  export const acids = [
+    "HCl", // Hydrochloric acid
+    "H2SO4", // Sulfuric acid
+    "HNO3", // Nitric acid
+    "H3PO4", // Phosphoric acid
+    "CH3COOH", // Acetic acid
+    "H2CO3", // Carbonic acid
+    "HF", // Hydrofluoric acid
+    "HBr", // Hydrobromic acid
+    "HI", // Hydroiodic acid
+    "HClO", // Hypochlorous acid
+    "HClO2", // Chlorous acid
+    "HClO3", // Chloric acid
+    "HClO4", // Perchloric acid
+    "H2S", // Hydrosulfuric acid
+  ];
+  
+  export const bases = [
+    "NaOH", // Sodium hydroxide
+    "KOH", // Potassium hydroxide
+    "Ca(OH)2", // Calcium hydroxide
+    "Mg(OH)2", // Magnesium hydroxide
+    "NH3", // Ammonia
+    "Al(OH)3", // Aluminum hydroxide
+    "Fe(OH)3", // Iron (III) hydroxide
+    "NH4OH", // Ammonium hydroxide
+    "Ba(OH)2", // Barium hydroxide
+    "Sr(OH)2", // Strontium hydroxide
+    "LiOH", // Lithium hydroxide
+  ];
