+++ chemistryData/constants.js
@@ -1,0 +1,6 @@
+export const physicalConstants = {
+    R: 0.0821, // Ideal gas constant in L·atm/(mol·K)
+    standardTemperature: 273.15, // Kelvin
+    standardPressure: 1, // atm
+    molarVolume: 22.4, // L/mol at STP
+  };
