+++ chemistryData/hydrocarbons.js
@@ -1,0 +1,43 @@
+export const hydrocarbonWeights = {
+    CH4: 16.04, // Methane
+    C2H6: 30.07, // Ethane
+    C3H8: 44.1, // Propane
+    C4H10: 58.12, // Butane
+    C5H12: 72.15, // Pentane
+    C6H14: 86.18, // Hexane
+    C7H16: 100.21, // Heptane
+    C8H18: 114.22, // Octane
+    C9H20: 128.25, // Nonane
+    C10H22: 142.29, // Decane
+  
+    C2H2: 26.04, // Acetylene (Ethyne)
+    C2H4: 28.05, // Ethylene (Ethene)
+    C3H6: 42.08, // Propylene (Propene)
+    C4H8: 56.11, // Butylene (Butene)
+    C5H10: 70.13, // Pentene
+    C6H12: 84.16, // Hexene
+    C7H14: 98.19, // Heptene
+    C8H16: 112.21, // Octene
+    C9H18: 126.24, // Nonene
+    C10H20: 140.27, // Decene
+  
+    C6H6: 78.11, // Benzene
+    C7H8: 92.14, // Toluene
+    C8H10: 106.17, // Xylene
+    C9H12: 120.19, // Trimethylbenzene
+    C10H14: 134.22, // Tetramethylbenzene
+  
+    C6H10: 82.15, // Cyclohexene
+    C6H12: 84.16, // Cyclohexane
+    C7H14: 98.19, // Cycloheptane
+    C8H16: 112.21, // Cyclooctane
+  
+    C10H8: 128.17, // Naphthalene
+    C14H10: 178.23, // Anthracene
+    C14H10: 178.23, // Phenanthrene
+    C16H10: 202.26, // Pyrene
+    C18H12: 228.29, // Chrysene
+  
+    C20H12: 252.31, // Benzo[a]pyrene
+    C24H12: 300.36, // Coronene
+  };
