+++ chemistryData/redoxAgents.js
@@ -1,0 +1,62 @@
+  export const oxidizingAgents = [
+    "O2", // Oxygen
+    "O3", // Ozone
+    "H2O2", // Hydrogen Peroxide
+    "KMnO4", // Potassium Permanganate
+    "HNO3", // Nitric Acid
+    "K2Cr2O7", // Potassium Dichromate
+    "CrO3", // Chromium Trioxide
+    "H2SO4", // Sulfuric Acid (concentrated)
+    "Cl2", // Chlorine
+    "Br2", // Bromine
+    "I2", // Iodine
+    "FeCl3", // Ferric Chloride
+    "CuSO4", // Copper(II) Sulfate
+    "MnO2", // Manganese Dioxide
+    "AgNO3", // Silver Nitrate
+    "NaOCl", // Sodium Hypochlorite (Bleach)
+    "PbO2", // Lead(IV) Oxide
+    "N2O", // Nitrous Oxide
+    "NO2", // Nitrogen Dioxide
+    "HClO", // Hypochlorous Acid
+    "ClO2", // Chlorine Dioxide
+    "HBrO", // Hypobromous Acid
+    "HIO3", // Iodic Acid
+    "XeF4", // Xenon Tetrafluoride
+    "C2H5NO2", // Glycine Nitrate
+  ];
+  
+  export const reducingAgents = [
+    "H2", // Hydrogen Gas
+    "C", // Carbon (Graphite)
+    "CO", // Carbon Monoxide
+    "CH4", // Methane
+    "C2H6", // Ethane
+    "C6H6", // Benzene
+    "Fe", // Iron
+    "Zn", // Zinc
+    "Al", // Aluminum
+    "Mg", // Magnesium
+    "Na", // Sodium
+    "K", // Potassium
+    "Li", // Lithium
+    "Ca", // Calcium
+    "Ba", // Barium
+    "Sn", // Tin
+    "Pb", // Lead
+    "NH3", // Ammonia
+    "H2S", // Hydrogen Sulfide
+    "NaBH4", // Sodium Borohydride
+    "LiAlH4", // Lithium Aluminum Hydride
+    "SO2", // Sulfur Dioxide
+    "FeSO4", // Iron(II) Sulfate
+    "TiCl3", // Titanium(III) Chloride
+    "Cu", // Copper
+    "Cr2O3", // Chromium(III) Oxide
+    "AsH3", // Arsine
+    "SbH3", // Stibine
+    "PH3", // Phosphine
+  ];
+  
+  // Common combustion fuels
+  export const validCombustionFuels = ["CH4", "C2H6", "C6H6"];
